Pipe: refactor package structure (node-commons)
